backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
watchdog==4.0.2
